converting slug page to typescript
